// src/enums/common.enum.ts
export enum EJobTitle {
  SoftwareEngineer = 'Software Engineer',
  DataScientist = 'Data Scientist',
  OperationsManager = 'Operations Manager',
  ProjectManager = 'Project Manager',
}

export enum EDepartment {
  Finance = 'Finance',
  HumanResources = 'Human Resources',
  Marketing = 'Marketing',
  Operations = 'Operations',
  Sales = 'Sales',
  SoftwareDevelopment = 'Software Development',
}